+++ examples/sync_demo.py
@@ -4,33 +4,12 @@ from mli import SyncMLIClient
 def sync_demo_candle_stable_lm():
     sync_client = SyncMLIClient('http://127.0.0.1:5000')
 
-    print(sync_client.text(**{
-        "engine": "candle",
-        "kind": "stable-lm",
-        "model_id": "lmz/candle-stablelm-3b-4e1t",
-        "sample_len": 512,
-        "prompt": "Building a website can be done in 10 simple steps:\nStep 1:"
-    }))
-
-    print(sync_client.chat(**{
-        "engine": "candle",
-        "kind": "stable-lm",
-        "model_id": "lmz/candle-stablelm-3b-4e1t",
-        "sample_len": 512,
-        "messages": [
-            {"role": "system", "content": "You are a helpful assistant."},
-            {"role": "user", "content": "I need help building a website."},
-            {"role": "assistant", "content": "Sure, let me know what and hwo do you need it built."},
-            {"role": "user", "content": "Building a website can be done in 10 simple steps. Explain step by step."}
-        ]
-    }))
-
     for chunk in sync_client.iter_text(**{
         "engine": "candle",
         "kind": "stable-lm",
         "model_id": "lmz/candle-stablelm-3b-4e1t",
-        "sample_len": 512,
-        "prompt": "Building a website can be done in 10 simple steps:\nStep 1:"
+        "sample_len": 8 * 1024,
+        "prompt": "Building a perfect e-commerce website in 1234 simple steps:\nStep 1:"
     }):
         print(chunk, sep='', end='', flush=True)
 
@@ -38,12 +17,12 @@ def sync_demo_candle_stable_lm():
         "engine": "candle",
         "kind": "stable-lm",
         "model_id": "lmz/candle-stablelm-3b-4e1t",
-        "sample_len": 512,
+        "sample_len": 8 * 1024,
         "messages": [
             {"role": "system", "content": "You are a helpful assistant."},
             {"role": "user", "content": "I need help building a website."},
             {"role": "assistant", "content": "Sure, let me know what and hwo do you need it built."},
-            {"role": "user", "content": "Building a website can be done in 10 simple steps. Explain step by step."}
+            {"role": "user", "content": "Building a perfect e-commerce website in 1234 simple steps. Explain step by step."}
         ]
     }):
         print(chunk, sep='', end='', flush=True)
@@ -55,8 +34,38 @@ def sync_demo_candle_llama():
     for chunk in sync_client.iter_text(**{
         "engine": "candle",
         "kind": "quantized",
-        "model": "mistral-7b-v0.1.Q4_K_M.gguf",
-        # "model": "yarn-llama-2-7b-128k.Q4_K_M.gguf",
+        # "model": "mistral-7b-v0.1.Q4_K_M.gguf", # bad
+        "model": "orca-2-7b.Q4_K_M.gguf", # good
+        # "model": "yarn-llama-2-7b-128k.Q4_K_M.gguf", # good
+        # "model": "rocket-3b.Q4_K_M.gguf", # bad
+        "sample_len": 8 * 1024,
+        "prompt": "Building a perfect e-commerce website in 1234 simple steps:\nStep 1:"
+    }):
+        print(chunk, sep='', end='', flush=True)
+
+
+def sync_demo_llama_cpp_main_llama2():
+    sync_client = SyncMLIClient('http://127.0.0.1:5000')
+
+    for chunk in sync_client.iter_text(**{
+        "engine": "llama.cpp",
+        "kind": "main",
+        "n_gpu_layers": 35,
+        "model": "orca-2-7b.Q4_K_M.gguf",
+        "sample_len": 8 * 1024,
+        "prompt": "Building a perfect e-commerce website in 1234 simple steps:\nStep 1:"
+    }):
+        print(chunk, sep='', end='', flush=True)
+
+
+def sync_demo_llama_cpp_main_stablelm():
+    sync_client = SyncMLIClient('http://127.0.0.1:5000')
+
+    for chunk in sync_client.iter_text(**{
+        "engine": "llama.cpp",
+        "kind": "main",
+        "n_gpu_layers": 35,
+        "model": "rocket-3b.Q4_K_M.gguf",
         "sample_len": 8 * 1024,
         "prompt": "Building a perfect e-commerce website in 1234 simple steps:\nStep 1:"
     }):
@@ -64,5 +73,7 @@ def sync_demo_candle_llama():
 
 
 if __name__ == '__main__':
-    # sync_demo_candle_stable_lm()
+    sync_demo_candle_stable_lm()
     sync_demo_candle_llama()
+    sync_demo_llama_cpp_main_llama2()
+    sync_demo_llama_cpp_main_stablelm()